+++ oidc-idp/src/main/webapp/WEB-INF/user-context.xml
@@ -111,7 +111,7 @@
 	</bean>
 
 	<!-- defines our own user info service -->
-	<bean id="userInfoService" primary="true" class="cz.muni.ics.oidc.server.PerunUserInfoService">
+	<bean id="userInfoService" primary="true" class="cz.muni.ics.oidc.server.userInfo.PerunUserInfoService">
 		<property name="perunConnector" ref="perunConnector"/>
 		<property name="subAttribute" value="${attribute.openid.sub}"/>
 		<property name="preferredUsernameAttribute" value="${attribute.profile.preferred_username}"/>
